refactor: follow express naming conventions for files
